gogo_proto_library: allow passing deps to the go_library
This is used to depend on the well-known types (WKT), for example:

    proto_library(
        name = "protobuf",
        srcs = ["wkt.proto"],
        deps = [
            "@com_github_protocolbuffers_protobuf//:any_proto",
            "@com_github_protocolbuffers_protobuf//:wrappers_proto",
        ],
    )

    gogo_proto_library(
        name = "protolib",
        srcs = [":protobuf"],
        importpath = "github.com/pubref/rules_protobuf/examples/wkt/go/protolib",
        deps = [
            "@com_github_golang_protobuf//ptypes/any:go_default_library",
            "@com_github_golang_protobuf//ptypes/wrappers:go_default_library",
        ],
    )

The ideal names for these properties was not obvious. From a user's point of
view, I think that "srcs" makes sense as the list of proto_libraries, and deps
is the list of Go libraries, but this is not clear-cut. I also considered
"proto_libraries" and "go_libraries" to make it more explicit, but that seemed
counter to Bazel's naming conventions.